Put mesh types into their own files
