+++ src/constants.h
@@ -1,0 +1,6 @@
+#pragma once
+
+constexpr bool USE_SHADER = true;
+constexpr int MAX_KELVINLETS = 40;
+constexpr int FRAME_RATE = 30;
+constexpr bool SAVE_SCREENSHOTS = false;
